+++ doc/about/changes.md
@@ -16,6 +16,7 @@ in progress
   iterable in the appropriate build target
 - `info()`, `debug()`, and `warning()` now take a variable number of arguments
   to print
+- Add support for setting the entry point for native binaries (C, C++, etc)
 
 ### Breaking changes
 - Drop support for Python 2